Imagique/requirements.txt
numpy==1.24.0
opencv_python==4.6.0.66
pandas==1.5.2
Pillow==9.4.0
PyQt5==5.15.9
QtRangeSlider==0.1.5
